requirement.txt
websocket-client==1.6.4
kafka-python==2.0.2
pyspark==3.2.0
psycopg2-binary==2.9.9
pyyaml==6.0.1
python-dotenv==1.0.0